Reviee on Crud Functionality

+++ CRud/category.js
@@ -8,7 +8,10 @@ const categoryForm = document.getElementById('category-form');
 const CATEGORY_KEY = 'categories';
 // Read category data from local storage
 let categories = JSON.parse(localStorage.getItem(CATEGORY_KEY)) || [];
+/*
 // Display each category as a clickable element
+{{{{       U-  Operation(Update) Number tow               }}}}
+*/
 function updateCategoryList() {
 categoryList.innerHTML = '';
 categories.forEach(category => {
@@ -18,17 +21,18 @@ categoryElement.textContent = category;
 categoryList.appendChild(categoryElement);
 });
 }
-
 updateCategoryList();
 
+
+/*
 // Create new category
+{{{{              C-   Operation(create) Number one                       }}}}
+*/
 categoryForm.addEventListener('submit', function(event) {
 event.preventDefault();
 const newCategory = document.getElementById('new-category').value.trim();
 if (!newCategory){
-// Show error message for empty category
 alert("some thing went wrong !!");
-// return;
 }
 if (categories.includes(newCategory)) {
 return;
@@ -39,34 +43,51 @@ updateCategoryList();
 categoryForm.reset();
 });
 
+/*
+Edit function 
+{{{{       U-  Operation(Edit) Number tow               }}}}
 // Edit category
-const category = document.getElementById("category-list").value;
-function editCategory(category) {
-const index  = categories.indexOf(category);
-const value_ = document.getElementById("new-category");
-value_.value = categories[index].category;
+*/
+function editCategory() {
+  const selectedCategory = categoryList.value;
+  const index = categories.indexOf(selectedCategory);
+  if (index !== -1) {
+  // document.getElementById("new-category").value =  selectedCategory;
+  // newValue = document.getElementById("new-category").value;
+  // const  = prompt('Enter the new value for the category:', selectedCategory);
+    const newValue = prompt('Enter the new value for the category:', selectedCategory);
+    if (newValue){
+      categories[index] = newValue.trim();
+      localStorage.setItem(CATEGORY_KEY, JSON.stringify(categories));
+      updateCategoryList();
+    }
+  }
 }
 // Edit 
 const editButton = document.getElementById('buttonEdit');
-editButton.addEventListener('click',() => editCategory(category));
+editButton.addEventListener('click',() => editCategory());
 
+
+/*
 // Delete category
-function deleteCategory(category){
-const index = categories.indexOf(category);
-if (index !== -1) {
-categories.splice(index, 1);
-localStorage.setItem(CATEGORY_KEY, JSON.stringify(categories));
-updateCategoryList();
+Clicking "Delete" on a product item removes it from the categories array and updates local storage,
+ reflecting the deletion in the displayed list.
+{{{{       R-  Operation(Delete) Number four               }}}}
+*/
+function deleteCategory(){
+  const selectedCategory = categoryList.value;
+  const index = categories.indexOf(selectedCategory);
+  if (index !== -1) {
+    categories.splice(index, 1);
+    localStorage.setItem(CATEGORY_KEY, JSON.stringify(categories));
+    updateCategoryList();
+  }
 }
-}
-/// 
-
 
 
 function searchProdctCategory(){
     searchValue = document.getElementById("Search");
     var productContainer = JSON.parse(localStorage.getItem('allproduct'))
-
     var trs='';
     for(var i =0 ; i<productContainer.length;i++){
       if(productContainer[i].category.toLowerCase().includes(searchValue.value.toLowerCase()) )
